requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
pypng==0.20220715.0
regex==2023.3.23
requests==2.28.2
tiktoken==0.3.3
urllib3==1.26.15
websockets==11.0.2
